compiler: add the `nkNimNodeLit` node kind
## Summary
AST that represents the content of a `NimNode` tree was not detectable
as such without inspecting the surrounding AST. This complicated logic
inspecting the bodies of macros or `.compileTime` procedures, as it had
to specifically ignore the symbol/symbol-choice nodes part of the AST
for `mNBindSym` magic calls.

The existence of `NimNode` literals inside normal AST is now made
explicit and also detectable without further context via a dedicated
node kind, also preventing it from going unnoticed if erroneously
present in places where it shouldn't be.

## Details
- use `nkNimNodeLit` nodes when deserializing `NimNode`s in
  `vmcompilerserdes`
- static `bindSym` calls are now expanded into a `nkNimNodeLit` during
  semantic analysis
- don't use an extra node for providing source information to
  `opBindSym` -- the source information is already provided by the VM
- the arguments to dynamic `bindSym` calls are now considered during
  constant folding
- remove the `opcNBindSym` VM instruction
- remove the now obsolete `rvmInvalidBindSym` report
- propagate error nodes in `semBindSym`
- include `semBindSym` in compiler tracing

## Changes to typed AST
```nim
...
# inside a compile-time evaluation context (macro, `.compileTime` procedure, or `static` expression/statement)

bindSym("s")
# is now represented in the AST as:
# `NimNodeLit(Sym "s")`
# instead of:
# `Call(Sym "bindSym", Sym "s")`

let ident = ...
bindSym(ident)
# is now represented in the AST as:
# `Call(Sym "bindSym", Sym "ident", Sym "brClosed", IntLit ...)
# instead of:
# `Call(Sym "bindSym", Sym "ident", Sym "brClosed", ..., IntLit ...)
```