tests: disable the `realtimeGC` test
## Summary

The test depends on the GC not running at the wrong time, which can
happen if allocation behaviour changes (as is the case with
https://github.com/nim-works/nimskull/pull/488)

This is caused by the GC instance used for the shared library being
initialized with an incorrect stack-bottom value, leading to valid
stack cells located below the wrong bottom not being considered during
stack scanning. If a heap location is only referenced from said stack
cells, it is garbage collected, making further access to it invalid
and potentially causing crashes due to heap corruption.

For programs using both shared libraries and `refc`, the `nimrtl` (Nim
run-time library) has to be used in order for there to only be a single
GC instance per thread, but the test explicitly states that it tests
the realtime GC *without* using the `nimrtl`. This can't work, and only
did so in the past by accident.

The test is only disabled instead of removed, so that it can be, if
desired, corrected in the future without losing the associated
history.