vm: clean up the external interface
## Summary
Invoking the VM was previously done through an `execute` procedure that
was mainly meant for use by the compiler, as it had on-demand procedure
code-gen (via `vmjit`) and VM-data deserialization built in. The final
goal is to completely decouple the VM from the rest of compiler, and
breaking this point of coupling is another step towards this goal.

To encapsulate the state of a single thread of execution, the `VmThread`
object is introduced. `initVmThread` creates a new instance; `execute`
runs a thread until it yields; and `dispose` cleans up the resources
used by the thread and destroys it.

`ExecResult` is renamed to `YieldReason` and VM errors are now also
reported through the `YieldReason` object. The previous `vmjit`
integration as well the VM-data -> ``PNode`` translation are moved into
a procedure in the `compilerbridge` module that is implemented on top of
the `VmThread` primitive.

In addition, the `opcQuit` instruction no longer implements the quit
logic itself, but yields execution with a dedicated reason, leaving how
a `quit` call is to be handled up the VM's callsite. This is done as
experiment to see if a similar, but more generalized mechanism could
in the future be used to replace the current VM callbacks.

As a direct consequence, the `vm` no longer imports `vmjit` (which
imported `vmgen`).

## Details
- adjust the `vmrunner` to work with the VM API
- export `createStackTrace` and use a `VmThread` for providing the PC
  and stack frame position
- document issues with `opcGorge`, `opcEcho`, and the way ``quit``
  works at compile-time