compiler: improve `for`-loop tuple unpacking
## Summary
- fix an issue with `semForVar` where types were modified, leading to
  unrelated valid code being rejected by semantic analysis (see
   `tforloop_unpacking_bug.nim`)
- always use a full copy instead of a shallow copy for `for` variables
- refactor `semForVar`
- refactor `transformYield`
- make `semForVar` `nkError` aware

## Details
- fix the `nkHiddenAddr` nodes produced by `transformYield` having
  the wrong type (the type of the accessed tuple was used)
- split up `semForVar` into multiple procedures in order to reduce code
  duplication
- add debug-tracing calls to the `for`-loop related procedures
- improve the documentation of the refactored code
- add tests for `for`-loop tuple unpacking

Unconditionally using `nkFastAsgn` (i.e. a shallow assignment) for local
variables is unsound without proving that doing so is safe (which
did not happen). Figuring out where to use shallow copies should be
left to the cursor inference step.

The `nkFastAsgn` change negatively impacts performance of the
compiled code when using `--gc:refc` (or other legacy GCs), because a
full copy is materialized where it previously wasn't. This is not the
case for `--gc:arc|orc`, as `nkFastAsgn` was/is treated the same as
`nkAsgn` (i.e. as a full copy) for locations with destructors.