vmbackend: fix crash related to `nkEmpty` nodes
## Summary
During both serialization and deserialization of `PNode`s used for the
VM's RTTI, `nkEmpty` nodes were not explicitly handled, leading to them
being treated as nodes with children. This caused a `FieldDefect` when
trying to access their (inaccessible) `sons` field.

The problem rendered a large amount of programs unusable with the VM
backend, as either the compiler or the runner would crash when RTTI that
contained empty nodes somewhere was used.

## Details
- add `nkError` to the `nkWithoutSons` set
- add the `nkWithSons` set
- make `case` statements in `storeNode` and `loadNode` exhaustive.
  `nkNone` and `nkError` are now also detected (and rejected)
- add a test to make sure `nkEmpty` nodes are handled correctly