semexprs: refactor 'var' parameter argument fixup
## Summary
This is mostly a refactoring and clean up, but it also adjusts the "is
address taken" analysis to no longer ignore l-value conversions and
statement list expressions. The operand to an `addr` operation now also
uses the same (more correct) analysis as the arguments to `var`
parameters.

`jsgen` depends on the `sfAddrTaken` flag in order to know whether a
location needs to support being modified via a "base + index" (the JS
back-end's way of implementing pointers), so this change fixes a
JS code-gen issue. Only in theory however, as the JS code-generator
implements the relevant bits incorrectly -- although in slightly
different way, wrong code is still generated.

## Details
- rename `analyseIfAddressTakenInCall` to `fixVarArgumentsAndAnalyse`,
  refactor and document it, and make it `nkError` aware
- remove the mutability check from `newHiddenAddrTaken` -- this is now
  the responsibility of the callsite
- move the "is address taken" analysis to a separate procedure
- move the `FakeVarParams` constant to the `ast_query` module
- make `semAddrArg` `nkError` aware, fixing `addr x` reporting an
  `expression has no address` error if expression `x` had an error
- fix the "has no address" error reported for  `addr` always suggesting
  the usage of `unsafeAddr`, even if it was already used
- remove leftover and unused `unsafeAddr`-related logic