vmgen: move `mIs` folding into `semfold`
## Summary
The logic is copied verbatim, and while it is still incorrect, it is now
applied earlier and no longer reaches into the code-generator. In other
words, the issue (type expression that is erroneously not materilalized)
is worked around at a point that is closer (in the time sense) to the
origin of the problem than it was previously.