tables: fix multiple `OrderedTable` bugs
## Summary
The fixed issues:
- for `OrderedTable`s with a single item, the iterators would get stuck
  in an infinite loop in some cases
- items would sometimes not be yielded when using any of the iterators
- comparing two `OrderedTable` would sometimes ignore items

These only happened for `OrdererdTable` not explicitly initialized via
`initOrderedTable` - explicitly initialized ones were not affected.

## The problem
The underlying issue was the same in each case. An `OrderedTable` uses
an intrusive linked list internally. Each slot stores a `next` pointer
and the table itself the head (`first`) and tail (`last`) pointer.

Since the pointers are indices into the internal data `seq` and not
memory adresses, the default value (i.e. 0) does not represent "none"
(because the 0-th `seq` item is a valid item), meaning that the
pointers have to be explicitly initialized to the "none" state (i.e.
`-1`). This happens when using `initOrderedTable`, and at one point did
also happen when inserting into a default-initialized table, but the
latter behaviour got removed as part of
#22e84d869efae228dc2859a9fe1fe303b6bd58f7 (likely by accident).

When inserting an item into a default initialized `OrderedTable` and
the selected slot is slot `0`, the slot's `next` pointer was set to
point to the slot itself, due to the `last` pointer incorrectly
pointing to slot `0` at that point. If any of the iterators were used
directly after this insertion, they would loop forever (because of the
self-cycle).

When inserting an item into a default initialized `OrderedTable` and the
selected slot is not `0`, the `next` pointer of the unfilled slot `0`
(which is treated as the head of the linked list) is set to the index
of the newly filled slot. As most of the `OrderedTable` implementation
checks for and skips unfilled slots, this doesn't cause any immediate
problems, but once the slot at position `0` becomes used (before the
table is first enlarged), all previously inserted items are effectively
cut off. They're still part of the table, but are no longer returned
by any of the iterators nor considered during comparisons or when using
`sort`.

## Details
- add an overload of the `initImpl` template that correctly sets the
  `first` and `last` pointer to `-1` as part when initializing the
  `OrderedTable` (either explicitly or lazily)
- remove the `isFilled` guards from logic that traverses the linked
  list - they're unnecessary
- remove the `rawGetKnownHC`, `rawGetDeep`, and `rawGet` overloads for
  `OrderedTable`. They were either unused or the same as the more
  general implementation