introduce vm syscall - echo
This is the first 'syscall' for the VM. Meaning the VM will yield
control to the caller with the reason `yrkEcho`. The caller, the
"system" in this case, is responsible for fullfilling the expectations
of the echo, whatever that mean in the particular context, then resuming
execution, if it chooses to. Baby CPS (Continuation Passing Style) for
the win!

It's pretty simple right now, but it already drops more report related
dependencies from the VM. A bit more of this and we'll have excised
much of it in no time. :D