vmgen: small refactoring
## Summary
- refactor and improve the code in multiple places, with a focus on
  removing usages of `TGenFlag`
- don't materialize a full copy when passing locals to immutable
  parameters

This is as precursor to reworking address handling in the VM and
`vmgen`.

## Details
- split `genRdVar` into `genSym` and `genSymAddr`
- don't propagate the `gfNodeAddr` flag; handle the operands to
  `nkAddr`/`nkHiddenAddr` directly in `genAddr`
- remove leftover type expression handling
- clean up `genArrAccess`
- deduplicate object access logic
- remove the unused `genUnaryStmt` procedure

Use the "direct assignment" path for all locals -- `skTemp` symbols no
longer reach `vmgen` anymore, so the path was taken for all locals
anyway. This allows for always using a register copy in `genSym`, which
in turn removes the need for the `gfIsParam` flag and greatly improves
the run-time efficiency of passing complex locals to immutable
parameters.

The `tslow_tables.nim` test now compiles in ~3 instead of ~17.5
seconds.