governance: contributor guidelines change approval
Defined how changes, especially ones with larger impact are approved,
along with how tie breaking is handled.

Long story short: if there if the cost of changing a decision in a PR is
small and the contributor has a track record of consistently improving
things, then core members shouldn't needlessly
block changes. Finally, in the case of
impasses or matters of taste and direction it's my (Saem's) call.

In addition, the docs was
hard to follow and contained lots of outdated information. Much of it
has been pared back and layout reworked more towards the logical flow
for a contributor.

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>