vm: prepare for the `var`/address rework
## Summary
Introduce the `opcWrLoc` operation, which makes the `writeLoc`
procedure available at the instruction level. This is a prerequisite for
the rework of how `var` parameters and views are implemented by
`vmgen`.

The introduction of `opcWrLoc` allows for generating more efficient
code for `new` and `newSeq`, also fixing a double-evaluation bug with
their arguments.

## Details
- add `opcWrLoc`. It is different from `opcAsgnComplex` in that the
  former is used for writing/assigning to a *location*, while the latter
  is used for assigning a value to a *register*
- don't generate a patch assignment for `opcSetLenStr` or
  `opcSetLengthSeq`. The operations already mutate the correct location