cgen: make `.global.`s work more according to spec
For the C target, globals defined inside procedures that are resource-
like (i.e. have a user-defined or lifted destructor) are now
initialized in a module's pre-init procedure, instead of each time
control-flow reaches the definition. This mirrors what already happened
for non-resource-like globals.

In addition, globals defined in `inline` procedures no longer cause
linking errors because of duplicated definitions in the generated C
code.

The globals are extracted before translating the AST to MIR code,
making the workaround in `mirgen` (that didn't work anyway) obsolete.
In order to work towards unifying the architecture of the back-ends,
`vmbackend` now also makes use of the pre-extraction, no longer
requiring extra logic in the code-generator.

- the destructor injection pass is not run for the initializer
  expression of globals defined inside procedures
- `.global.`s on the JS target still don't work according to
  specification