bootstrap the compiler and tools with ORC
Summary
=======

All tools built as part of `koch tools` as well as the second and third
iteration of bootrapping now use ORC as the memory management
strategy. The first iteration does not, as the current csources compiler
fails to compile the compiler.

This is the first step towards making ORC the default and removing all
legacy GC options.

Details
=======

The tools that are now built with `--gc:orc` are:
- `nimsuggest`
- `nimgrep`
- `vccexe`
- `vmrunner`
- `nim_dbg`
- `testament`

As a side-effect of the change, the compiler and the VM (both built-in
and standalone) are now significantly faster. For compiling itself,
the `--gc:orc`-built compiler is ~30% faster in `-d:release` mode than
the `--gc:refc` counterpart.

Other
-----

- fix a parameter passing aliasing violation in `semTry` that lead to
  memory corruption