vm: separate argument processing from `opcExpandToAst`
The operation now expects its operands to be `NimNode`s already. A new
instruction (`opcDataToAst`) is introduced for creating the AST
representation of VM data.

While also simplifying the VM a bit, the main reason behind the change is
to not having to provide the full AST of the template call expression,
as doing so is not possible when the code generator no longer operates
on `PNode` AST.