backends: attach procedure-level globals to the procedure
This fixes a severe layering violation, namely that procedure processing
mutated module-related state. The code for initializing all procedure-level
globals is now stored with a `Procedure`, which allows its consumer to
subject it to further processing.

In addition, MIR passes (currently only destructor injection) are now also
applied to this fragment, fixing a pre-existing issue.