vm: support the return-value optimization
Summary
-------

Allow (non-magic) procedures to modify the destination location
directly if the result of the call is immediately assigned to
something. While the optimization is never actually made use of right
now (a temporary is always used), this change makes it possible to do
so in the future.

Details
-------

Instead of the VM allocating the location as part of executing
`IndCallAsgn`, `vmgen` now emits bytecode for doing so at the callsite
(if the destination not a valid location).

For values that fit into registers, the result register is setup at the
start of a procedure via `LdNullReg` (nothing is emitted at the callsite
in this case).

While supporting the return-value optimization is the main goal, the
change is also a step towards:
1. allocating locations in a single place (i.e. via `LdNull`), which is
  a preparation for removing the `rkLocation` register kind
2. not having to store the return type in the function-table entry