vmgen: support the return-value optimization for magics
Summary
=======

Allow *magic* procedures to modify the destination location directly
if the result of the call is immediately assigned to something. The
optimization is never actually made use of right now.

This is another step towards removing location setup (outside of
`LdNull`) from the VM -- the only usage sites remaining are `opcConv`
and the logic for closure calls.

Details
=======

- support in-place object construction without going through a
  temporary (optimization is not made use of right now)
- don't allocate unnecessary temporaries for `mNccInc`, `mNcsAdd`, and
  `mNcsIncl`
- `withTemp` creates full temporaries now