types: fix all non-primitve types being treated as cyclic
Summary
=======

Due to a logic bug with `canFormAcycle`, *all* non-primitive types (such
as tuples, objects, `array`s, `seq`s, etc.) not explicitly marked with
`.acyclic` were detected as being cyclic.

Since `canFormAcycle` is used to decide whether a type is relevant to
the cycle collector when using `--gc:orc`, all heap locations of
non-primitive type were tracked and scanned by the cycle collector, and
unnecessary `=trace` hooks lifted for them.

As a consequence of the above, moving `ref` values between threads or
holding references to a (non-cyclic) heap location from multiple threads
was effectively impossible (without crashing) when using `--gc:orc`.

Details
=======

The logic for detecting types that can potentially act as the root in a
reference cycle is rewritten from scratch, now properly handling
structural types and generic instantiations. The `.cursor` annotation
is also taken into account.

Since `canFormAcycle` is still used by logic related to the legacy it
is kept for now, in order to reduce the impact of the fix.

For testing the whether is potentially cyclic, the `isCyclical`
type-trait query procedure is added to `std/typetraits`.

Misc
----

- remove leftover debug code from `orc.nim`