document how `--gc:arc` and `--gc:orc` work
Summary
=======

Only a very short overview of both was previously provided, with the
details only available by inspecting the implementation in the compiler
and compiler run-time.

In order to make discussion surrounding both easier and more accessible,
a detailed description of how both ARC and ORC work is now provided in
`gc.rst`.

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>