compiler: replace `Rope`s with plain strings
Summary
=======

The usage of `Rope`s is blocking the switch to `--gc:orc` for the
compiler, as very long chains that can happen in some cases (such as
`const` arrays with a very large number of items) lead to stack
overflows on `Rope` destruction.

In addition, replacing `Rope`s with strings leads to a decrease in peak
memory consumption (when building the compiler itself) by 400 MiB --
execution time is not affected for a compiler built with `refc`.

Details
=======

- make `Rope` an alias for `string`
- remove the now-unused routines from `ropes.nim`
- replace `nil` with `"" where `Rope` where previously used

Misc
----

- fix the "Unused import" warnings with regarding `debugutils`