vmgen/vm: conversion-related improvements
Summary
=======

Refactor the conversion-related parts in `vmgen`, which fixes multiple
bugs and issues with code running in the VM:
- conversions where the source was of `sink` or `static` type always
  created a full copy
- lvalue-ness was not preserved for conversions to or from `distinct T`
  types was not always preserved
- assignments where the left-hand side is an lvalue conversion didn't
  update the destination location in some cases (such as for `ref`s or
  field assignments)

In addition, the VM access checks are now less strict: accessing a
`ref|ptr T` location with a handle of `ref|ptr X` is allowed, as long
as `X` is either a sub- or super-type of `T`.

Details
=======

Move the conversion logic for `ref`s and `objects` to a new instruction
(`opcObjConv`). Compared to the previously used `opcConv`, using a
dedicated conversion operator has the benefits that:
- only one type argument is required instead of two, meaning that the
  instruction can be encoded with two instead of three instruction
  words
- no full RTTI is required

For `opConv`:
- remove the `initLocReg` calls -- making sure that the destination
  location is set-up is now the responsibility of the code generator
- remove the catch-all branch

Other
-----

- fold `PointerLike(nil)` expressions into correctly typed
  `nkNilLit`s during `transf`

Misc
----

- move the rendering of VM type conversion diagnostics to
  `cli_reporter` by introducing the dedicated
  `vmEvtIllegalConvFromXToY` event/diagnostic
- store the formal and actual type for VM type conversion reports
  directly, instead of as a `seq[TypeMismatch]`
- add a `isNotNil` routine for `HeapSlotHandle`