cgen: fix IC-backend crash with `inline` procedures
Summary
=======

Using an `inline` procedure in a different module from where it was
defined lead to a compiler crash when the defining module was not
pending. As a workaround, procedure-level globals are generated into the
module where the parent procedure is first *used* instead of where it's
*defined*.

The above means that procedure-level globals are now initialized in a
different order when using symbol files (i.e. incremental compilation)
compared to when not.

Details
=======

The fundamental problem is that procedure-level globals are not attached
to their owning routines early enough and in a way that they can be
properly special-cased by the code generators.