internals: `useNodeIds` now tracks outgoing copies
When the compiler is compiled with `-d:useNodeIds` it used to output a
"comes from" debug message when a node was created by copy with an id
matching `ast.nodeIdToDebug`. Now, it will also output a "goes to"
message when a node is created from a node with the matching id.

Hopefully it's more useful for compiler developers.