vm: make `opcExpandToAst` more flexible
Summary
=======

The `ExpandToAst` operation was previously responsible for turning
each arguments' value into `PNode` AST in order for them to be passed to
`evalTemplate`. For providing the `PType` of each argument, the AST of
the original template call expression was passed as the first operand.

Since this approach would become a problem once `vmgen` no longer
operates on `PNode` AST, a different one is required: the
deserialization logic is moved to a dedicated operation (`opcDataToAst`)
taking the value and the destination type as input, with `vmgen`
emitting the instruction for arguments to template calls.

Details
=======

- add the `opcDataToAst` operation
- clean up the `opcExpandToAst` implementation
- adjust the logic for `mExpandToAst` and clean up / improve the code

Misc
----

- fix `regToNode` not initializing the output node's line information
  in all cases